+++ src/index.ts
@@ -1,4 +1,4 @@
 export * from './serverless-app';
+export * from './custom';
 export * from './extensions';
 export * from './main';
-